requirements.txt
pynput
python-telegram-bot